Adding test case and code for when motif contains rests
Previous test cases do not have any rests in them. Adding test cases to
check for this. Additional code was required to ensure we select a
non-rest note for deletion, and also to collapse adjacent rests into a
single rest with the same total duration. This may turn out to be
superfluous but seemed cleaner in my eyes.